requirements.txt
pyfluidsynth==1.3.4
numpy==2.4.2
scipy==1.17.1
soundfile==0.13.1
redis==7.3.0
pydantic==2.12.5